rename WalletConnectPluginEventMethod to WalletConnectPluginEventType
